move wrong_self_convention to its own module
